Splitted UI to 3 files. UIResources, UIWindows and UI
